feat(routine): UX, storage guards, split UI, backup confirm
- Toggle completions only on scheduled days; normalize habits on update
- Mark all habits for day; storage error event + import failure handling
- Notify key cleanup; remove tagScope pref; roadmap + buildHubCalendarEvents test
- Split RoutineApp into panels; tab a11y; routineDay URL; bulk mark + settings search

Made-with: Cursor
