+++ docs/hub-modules-roadmap.md
@@ -1,6 +1,6 @@
 # Hub: модулі на оку (рішення та ТЗ)
 
-Цей документ фіксує **пріоритет доменів** для розвитку Hub і **мінімальне ТЗ** для модуля «Звички», узгоджені з поточною архітектурою (`lazy`-модулі, `localStorage`, хаб у [`src/core/App.jsx`](../src/core/App.jsx)).
+Цей документ фіксує **пріоритет доменів** для розвитку Hub і узгоджений з поточною архітектурою (`lazy`-модулі, `localStorage`, хаб у [`src/core/App.jsx`](../src/core/App.jsx)).
 
 ---
 
@@ -16,58 +16,40 @@
 
 ---
 
-## 2. Модуль «Звички» (Habits) — мінімальне ТЗ
+## 2. Модуль «Рутина» (звички + Hub-календар)
+
+Фактична реалізація: [`src/modules/routine/RoutineApp.jsx`](../src/modules/routine/RoutineApp.jsx), сховище [`hub_routine_v1`](../src/modules/routine/lib/routineStorage.js) (не окремий `hub_habits_v1`).
 
 ### 2.1 Мета
 
-Простий трекер щоденних звичок: відмітка виконання по днях, **streak**, без сервера.
+Трекер звичок із відмітками по днях, **streak**, інтеграція плану **Фізрука** і планових підписок **Фініка** в одному календарі, без сервера.
 
-### 2.2 Дані (чернетка схеми)
+### 2.2 Дані (орієнтир)
 
-Зберігання в `localStorage`, один ключ на сховище, наприклад `hub_habits_v1`:
+- Один ключ `localStorage`: `hub_routine_v1` (звички, completions, теги, категорії, відтискання, нотатки до відміток, порядок звичок).
+- Резервна копія: експорт/імпорт JSON з модуля (окремий формат `hub-routine-backup`).
+- Дата — рядок **`YYYY-MM-DD`** у локальному календарі.
 
-```json
-{
-  "schemaVersion": 1,
-  "habits": [
-    {
-      "id": "uuid",
-      "name": "Вода 2 л",
-      "emoji": "💧",
-      "createdAt": "ISO-8601",
-      "archived": false
-    }
-  ],
-  "completions": {
-    "uuid": ["2026-04-13", "2026-04-12"]
-  }
-}
-```
+### 2.3 UI (реалізовано)
 
-- Дата — рядок **`YYYY-MM-DD`** у локальному календарі користувача.
-- Ідентифікатор звички — стабільний `id`; `completions[id]` — відсортований або ні масив дат (дедуплікат при записі).
-
-### 2.3 UI (мінімум)
-
-- Список активних звичок + кнопка «Додати».
-- На головному екрані модуля: **сітка поточного місяця** або горизонтальний ряд останніх 7–14 днів (як у Фізруку за складністю).
-- Тап по дню — toggle виконання для обраної звички (або long-press для вибору звички, якщо одна картка на день).
+- Вкладки **Календар** / **Рутина** (налаштування, CRUD звичок, бекап).
+- Зрізи часу (сьогодні, завтра, тиждень, місяць), тижневий ряд, місячна сітка, стрічка подій.
+- Параметр URL `routineDay=YYYY-MM-DD` відкриває зріз одного дня.
 
 ### 2.4 Інтеграція з Hub
 
-- Новий lazy-модуль: `src/modules/habits/HabitsApp.jsx` (або `HubHabits.jsx`).
-- У [`App.jsx`](../src/core/App.jsx): додати `habits` до `VALID_MODULES` і об’єкта вибору модуля (іконка + опис).
-- Експорт у повний бекап Hub (за аналогією з [`fizrukStorage`](../src/modules/fizruk/lib/fizrukStorage.js) / окремий `hubBackupKeys` у корені, якщо з’явиться спільний експорт).
+- Модуль підключено в [`App.jsx`](../src/core/App.jsx) як `module=routine`.
+- Єдиний повний бекап Hub (якщо з’явиться) — окремий майбутній крок; поки експорт Рутини з модуля.
 
-### 2.5 Не входить у v0
+### 2.5 Не входить у поточну версію
 
-- Спільний акаунт / синхронізація.
-- Push-нагадування (можна пізніше через той самий SW, що й для Фізрука).
-- Складна аналітика (лише streak + календар відміток).
+- Спільний акаунт / синхронізація між пристроями.
+- Push у фоні без обмежень браузера (можливий майбутній SW).
+- Складна аналітика (лише streak + події в календарі).
 
-### 2.6 Зв’язок з Фізруком (опційно пізніше)
+### 2.6 Зв’язок з Фізруком
 
-- Діплінк з дашборду Фізрука: «Звички» як доповнення до тренувань, без обов’язкового merge даних у v0.
+- План тренувань підтягується зі сховища Фізрука; перехід у модуль Фізрука зі стрічки.
 
 ---
 
@@ -75,11 +57,11 @@
 
 | Напрям | Примітка |
 |--------|----------|
-| Харчування lite | Після Habits або як підмодуль Фізрука (вага вже є в замірах). |
+| Харчування lite | Після розширення рутини або як підмодуль Фізрука (вага вже є в замірах). |
 | Сон / wellbeing | Розширення даних поруч із post-workout wellbeing у Фізруку або окремий екран. |
-| Єдиний інбокс нагадувань | Після того, як буде 2+ джерела подій (Фінік, Фізрук, Habits). |
+| Єдиний інбокс нагадувань | Після того, як буде 2+ джерела подій (Фінік, Фізрук, Рутина). |
 | Нотатки Markdown | Низький ризик технічно; середній пріоритет продукту. |
 
 ---
 
-*Оновлено для узгодження з планом «Модулі Hub на майбутнє».*
+*Оновлено під модуль «Рутина» та `hub_routine_v1`.*
